Refactor: extract inner static class
